Move form to its own module
