+++ src/style.css
@@ -1,3 +1,3 @@
-.projectModal {
+.project-form {
   display: none;
 }
