+++ src/ui-tasks.js
@@ -2,6 +2,7 @@ import { format } from "date-fns";
 import { todoFactory } from "./todo.js";
 import { projectFactory, projectProto } from "./project";
 import { updateProjectLS } from "./localstorage"
+import { form } from "./form"
 import "./style.css";
 
 const toDos = (project) => {
@@ -13,65 +14,8 @@ const toDos = (project) => {
     if (tasksList) {tasksList.remove()}
   }
 
-  // New task
-  const formContainer  = document.createElement("div")
-
-  const newTaskBtn = document.createElement("h3");
-  newTaskBtn.innerHTML = "New Task";
-
-  const taskForm = document.createElement("form");
-  taskForm.className = "taskModal";
-
-  const inputTitle = document.createElement("input");
-  inputTitle.setAttribute("type", "text");
-  inputTitle.setAttribute("name", "title");
-  inputTitle.setAttribute("placeholder", "Task name");
-
-  const inputDescription = document.createElement("input");
-  inputDescription.setAttribute("type", "text");
-  inputDescription.setAttribute("name", "description");
-  inputDescription.setAttribute("placeholder", "Description");
-
-  const inputPriority = document.createElement("select");
-  inputPriority.setAttribute("name", "priority");
-
-  const optionOne = document.createElement("option");
-  optionOne.setAttribute("value", "1");
-  optionOne.innerHTML = "High";
-
-  inputPriority.appendChild(optionOne);
-
-  const optionTwo = document.createElement("option");
-  optionTwo.setAttribute("value", "2");
-  optionTwo.innerHTML = "Medium";
-
-  inputPriority.appendChild(optionTwo);
-
-  const optionThree = document.createElement("option");
-  optionThree.setAttribute("value", "3");
-  optionThree.innerHTML = "Low";
-
-  inputPriority.appendChild(optionThree);
-
-  const inputDate = document.createElement("input");
-  inputDate.setAttribute("type", "date");
-  inputDate.setAttribute("name", "date");
-  inputDate.setAttribute("value", "yyyy-mm-dd");
-
-  const submitBtn = document.createElement("input");
-  submitBtn.setAttribute("type", "submit");
-  submitBtn.setAttribute("value", "OK");
-
-  taskForm.appendChild(inputTitle);
-  taskForm.appendChild(inputDescription);
-  taskForm.appendChild(inputPriority);
-  taskForm.appendChild(inputDate);
-  taskForm.appendChild(submitBtn);
-
-  formContainer.appendChild(newTaskBtn);
-  formContainer.appendChild(taskForm);
-
-  element.appendChild(formContainer);
+  let taskForm = form()
+  element.appendChild(taskForm);
 
   taskForm.addEventListener("submit", (e) => {
     e.preventDefault();
